fix Sync E2E tests
